backend/requirements.txt
fastapi
uvicorn
openai
pygithub
pydantic
python-dotenv
